Fix nonexhaustive match in linux
I briefly looked for the correct keycodes for MousewheelUp and Down, but couldn't find a canonical source that also matched the same keycode values for the existing entries. I don't have a Linux deployment to test on, so didn't feel comfortable guessing.

I think this is enough to get it compiling.